Add test-packages/example.addon.
This has a z3c.autoinclude.plugin entry point without a target.
We need a dummy option for this.
This scenario might not be needed after all.
But for now let's check how we handle this.
Currently, a plain load_package call finds our zcml.
A load_packages call with a specific target does not find it.
We can later opt to include our zcml in that case.