Drop support for PyPy.
Technically it may still work, but we no longer test it.  It is irrelevant for Plone.

Main reason for dropping it now, is that this needs an update to our gha setup.
The setup-python action [fails](https://github.com/plone/plone.autoinclude/actions/runs/13547585011/job/37862942913?pr=27):

Error: Invalid format of Python version for PyPy. Python version should be specified in format 'x.y'.