Fix expected values in tests.
In `example.ploneintegration`, the `example.addon` zcml gets loaded.
In `example.zopeintegration`, the `example.addon` and `example.multipleeps` zcml gets loaded.
This is because `example.{addon,multipleeps}/setup.py` has a strange entry point:

```
[z3c.autoinclude.plugin]
dummy = dummy
```

This was added to check if our code can handle unexpected entry point contents without breaking.
I don't really mind whether the zcml of these packages gets loaded or not, as this should not happen in practice.
But the new way of getting the available distributions and their entry points, apparently has this as side effect.